refactor: move loadInitialFile.service one level up
#2294
