+++ visualization/app/codeCharta/services/loadInitialFile/loadInitialFile.service.spec.ts
@@ -1,18 +1,18 @@
 import { TestBed } from "@angular/core/testing"
 import { HttpClient } from "@angular/common/http"
 import { MatDialog } from "@angular/material/dialog"
-import { CCFile, LayoutAlgorithm } from "../../../codeCharta.model"
-import { State } from "../../../state/angular-redux/state"
-import { GLOBAL_SETTINGS } from "../../../util/dataMocks"
-import { GlobalSettingsHelper } from "../../../util/globalSettingsHelper"
+import { CCFile, LayoutAlgorithm } from "../../codeCharta.model"
+import { State } from "../../state/angular-redux/state"
+import { GLOBAL_SETTINGS } from "../../util/dataMocks"
+import { GlobalSettingsHelper } from "../../util/globalSettingsHelper"
 import { LoadInitialFileService } from "./loadInitialFile.service"
-import { Store } from "../../../state/angular-redux/store"
-import { LoadFileService } from "../loadFile.service"
-import { ErrorDialogComponent } from "../../../ui/dialogs/errorDialog/errorDialog.component"
-import sample1 from "../../../assets/sample1.cc.json"
-import sample2 from "../../../assets/sample2.cc.json"
-import { FileSelectionState, FileState } from "../../../model/files/files"
-import { setFiles } from "../../../state/store/files/files.actions"
+import { Store } from "../../state/angular-redux/store"
+import { LoadFileService } from "../loadFile/loadFile.service"
+import { ErrorDialogComponent } from "../../ui/dialogs/errorDialog/errorDialog.component"
+import sample1 from "../../assets/sample1.cc.json"
+import sample2 from "../../assets/sample2.cc.json"
+import { FileSelectionState, FileState } from "../../model/files/files"
+import { setFiles } from "../../state/store/files/files.actions"
 
 describe("LoadInitialFileService", () => {
 	let loadInitialFileService: LoadInitialFileService
